Remove call to GenerateSeqNumberAndSubKey in krb5Token.go
The sequence number has already been generated in `types.NewAuthenticator`, and a subkey is not strictly required.

For instances where a subkey is required and depending on the use case (ie, when the Authenticator is generated through a call to `NewNegTokenInitKrb5`), it will need to be propagated back to the caller before the authenticator is serialized.