Add new WrapToken struct
New WrapToken struct and assorted functions to easily parse and serialize GSS-API per-message tokens.

This should enable the gokrb5 library to be easily usable for SASL exchanges where kerberos is used for authentication (ie, Hadoop clusters)

The WrapToken functions are tested, and this code currently works to authenticate against a hadoop namenode.