Respect the case of realm in the dns -> realm mapping configuration
While unusual, some Kerberos installations use lower case realm names. We
should respect what the configuration says.

Always making them upper case was making the session code fail to find the
existing session for the realm (eg. it was trying to find the session for
EXAMPLE.COM and not example.com) and issue a TGS-REQ as if we were trying to do
cross-realm authentication.

It makes sense to keep the canonicalization of the DNS name though.