service: Respect RequiredHostAddr when ticket contains client addresses
When the KDC is configured with noaddresses=false, the ticket will contain
client addresses it was issued for.

We still want to be able to disable the check on the service side though. It's
not unsual that the client address can't really be checked against, .eg SNAT
will effectly hide the client address we want to verify.

This commit changes the current behaviour: if the ticket contains client
addresses but RequireHostAddr is false (which is the default value),
ValidateAPREQ will now pass (it used to error out). The library user has to
explictly set RequireHostAddr to true for that check to happen.