Use 5 second timeout on initial connection to KDC
net.DialTCP() and net.DialUDP() don't allow specifying a timeout. This
means if the connection if a connection can't be established to the KDC
things hang until the OS timeout occurs which could be on the order of
minutes.

Instead, use net.DialTimeout() and specify a network (tcp, udp) and a
timeout of five seconds. This matches the read/write timeout set once
the connection has been established.