Redefine marshalNegTokenInit.ReqFlags as asn1.BitString
Change the defintion of ReqFlags from gssapi.ContextFlags (which is
itself a type alias for asn1.BitString) to an asn1.BitString directly.

This allows the asn1 type reflection to correctly identify the ReqFlags
field as a type 3 bitstring rather than a type 16 sequence, and will
parse the value of the ReqFlags field correctly. Without this change,
the parser will fail to interpret an `asn1:"explicit,optional,tag:1`
field without skipping the byte offset past the field, and such all
following tags will also fail to be parsed correctly.

Since the field has the same underlying datatype, it can be casted
to the relevant one in each direction to maintain the same behaviour
whilst also resolving the type reflection mis-identification.

Fixes #390